Seperate logic into `XcodeHashes` file.
